+++ src/windowLoadP.ts
@@ -1,0 +1,15 @@
+/**
+ * Returns a promise that resolves as soon as possible after the window is loaded.
+ * If the {@link https://developer.mozilla.org/en-US/docs/Web/API/Document/readyState `document.readyState`}
+ * is `'complete'` at call-time, the returned promise resolves immediately, otherwise it resolves upon
+ * the window load event.
+ *
+ * @return {Promise<void>}
+ */
+export const windowLoadP = () => new Promise<void>((resolve) => {
+    if (document.readyState === 'complete') {
+        resolve()
+    } else {
+        window.addEventListener('load', () => resolve())
+    }
+})